requirements.txt
pandas
numpy
scikit-learn
torch
transformers
stable-baselines3[extra]
requests
yfinance
newsapi-python
jupyter
matplotlib
seaborn
ta
gymnasium
quantstats
tqdm
PyYAML
python-dotenv